Refactor server into smaller files
